+++ src/main/java/com/indiraactive/fulfillmentplatform/dal/scriptRunAuditEntry/ScriptRunAuditEntryRepository.java
@@ -1,4 +1,4 @@
-package com.indiraactive.fulfillmentplatform.dal;
+package com.indiraactive.fulfillmentplatform.dal.scriptRunAuditEntry;
 
 import com.indiraactive.fulfillmentplatform.model.db.ScriptRunAuditEntry;
 import org.springframework.data.jpa.repository.JpaRepository;
