+++ src/main/java/com/indiraactive/fulfillmentplatform/dal/supplier/SupplierRepository.java
@@ -1,4 +1,4 @@
-package com.indiraactive.fulfillmentplatform.dal;
+package com.indiraactive.fulfillmentplatform.dal.supplier;
 
 import com.indiraactive.fulfillmentplatform.model.db.Supplier;
 import org.springframework.data.repository.CrudRepository;
