+++ src/main/java/com/indiraactive/fulfillmentplatform/manager/scheduledTask/ScheduledTaskManager.java
@@ -1,7 +1,6 @@
 package com.indiraactive.fulfillmentplatform.manager.scheduledTask;
 
-import com.indiraactive.fulfillmentplatform.dal.ScheduledTaskRepository;
-import com.indiraactive.fulfillmentplatform.dal.SupplierRepository;
+import com.indiraactive.fulfillmentplatform.dal.scheduledTask.ScheduledTaskRepository;
 import com.indiraactive.fulfillmentplatform.factory.ScheduledTaskJpaFactory;
 import com.indiraactive.fulfillmentplatform.model.ScheduledTask;
 import com.indiraactive.fulfillmentplatform.model.ScheduledTaskJpa;
