+++ src/main/java/com/indiraactive/fulfillmentplatform/service/AuditLoggerService.java
@@ -1,7 +1,6 @@
 package com.indiraactive.fulfillmentplatform.service;
 
-import com.indiraactive.fulfillmentplatform.dal.ScriptRunAuditEntryRepository;
-import com.indiraactive.fulfillmentplatform.factory.ScriptRunAuditEntryFactory;
+import com.indiraactive.fulfillmentplatform.dal.scriptRunAuditEntry.ScriptRunAuditEntryRepository;
 import com.indiraactive.fulfillmentplatform.model.db.ScriptRunAuditEntry;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
