+++ src/main/java/com/indiraactive/fulfillmentplatform/viewModel/ScheduleHistoryViewModel.java
@@ -1,6 +1,6 @@
 package com.indiraactive.fulfillmentplatform.viewModel;
 
-import com.indiraactive.fulfillmentplatform.dal.SupplierRepository;
+import com.indiraactive.fulfillmentplatform.dal.supplier.SupplierRepository;
 import com.indiraactive.fulfillmentplatform.model.ScheduledTaskHistoryModel;
 import com.indiraactive.fulfillmentplatform.model.ScheduledTaskJpa;
